Podcast Name implimentatio
This contains the code needed to impiment the getPodcastName function to make the general user experience better and remove any need for the user to enter a name.

The add function now adds the name to the .feeds file.
A new option in the main loop to show all the names and the function to get the names

This is all needed to run the remove podcast option